[bug] 추가정보 입력까지 api통신오류 모두 해결
