+++ src/pages/addition/selectedrole/components/optionsList.jsx
@@ -41,10 +41,10 @@ export const OptionsNumber = [
   { value: 22, label: 22 },
   { value: 23, label: 23 },
   { value: 24, label: 24 },
-  { value: 25, label: 25 },
-  { value: 26, label: 26 },
-  { value: 27, label: 27 },
-  { value: 28, label: 28 },
-  { value: 29, label: 29 },
-  { value: 30, label: 30 },
+  // { value: 25, label: 25 },
+  // { value: 26, label: 26 },
+  // { value: 27, label: 27 },
+  // { value: 28, label: 28 },
+  // { value: 29, label: 29 },
+  // { value: 30, label: 30 },
 ];
